fix: PersonalityEngine type cast + PK commit hash missing sender
PersonalityEngine.sol:
- Fix unsafe int8 cast: int16(int8(int256(uint256(current)))) could
  overflow when current > 127. Changed to int16(uint16(current))

PK commit-reveal (claw CLI):
- Fix critical bug: commit hash was keccak256(strategy || salt) but
  contract verifies keccak256(strategy || salt || msg.sender)
  This caused ALL reveals to fail with "Invalid reveal"
- Add wallet address to salt filename to prevent collision when
  same user controls both sides of a PK match (self-testing)

+++ contracts/core/PersonalityEngine.sol
@@ -186,7 +186,7 @@ contract PersonalityEngine is OwnableUpgradeable, UUPSUpgradeable {
     }
 
     function _clampPersonality(uint8 current, int8 delta) internal pure returns (uint8) {
-        int16 result = int16(int8(int256(uint256(current)))) + int16(delta);
+        int16 result = int16(uint16(current)) + int16(delta);
         if (result < 0) return 0;
         if (result > 100) return 100;
         return uint8(uint16(result));
